+++ NEWS.md
@@ -1,12 +1,13 @@
 ---
 title: "NEWS.md"
 author: "Simon Dedman"
-date: "2021-01-14"
+date: "2021-10-07"
 output: html_document
 ---
 
 # v1.5.1
-* More reporting info
+* More reporting info - Self_CV_Statistics.csv created with all outputs from all models for those 2 gbm.step objects
+* gbm.loop automatially checks for the presence of Report.csv's in numbered folders and doesn't run for that folder if so
 
 # v1.5.0
 * gbm.loop fixed for running gbm.auto & post-run-results-gatehring loops separately.